Move outbound networking config to new crate
