+++ crates/factor-outbound-http/src/wasi.rs
@@ -4,7 +4,8 @@ use anyhow::Context;
 use http::{header::HOST, Request};
 use http_body_util::BodyExt;
 use spin_factor_outbound_networking::{
-    BlockedNetworks, ComponentTlsClientConfigs, OutboundAllowedHosts, TlsClientConfig,
+    config::{allowed_hosts::OutboundAllowedHosts, blocked_networks::BlockedNetworks},
+    ComponentTlsClientConfigs, TlsClientConfig,
 };
 use spin_factors::{wasmtime::component::ResourceTable, RuntimeFactorsInstanceState};
 use tokio::{net::TcpStream, time::timeout};
