+++ crates/factor-outbound-mqtt/src/host.rs
@@ -2,7 +2,7 @@ use std::{sync::Arc, time::Duration};
 
 use anyhow::Result;
 use spin_core::{async_trait, wasmtime::component::Resource};
-use spin_factor_outbound_networking::OutboundAllowedHosts;
+use spin_factor_outbound_networking::config::allowed_hosts::OutboundAllowedHosts;
 use spin_world::v2::mqtt::{self as v2, Connection, Error, Qos};
 use tracing::{instrument, Level};
