+++ crates/factor-outbound-networking/Cargo.toml
@@ -9,20 +9,18 @@ anyhow = { workspace = true }
 futures-util = { workspace = true }
 http = { workspace = true }
 ip_network = "0.4.1"
-ip_network_table = "0.2.0"
 rustls = { workspace = true }
 rustls-pki-types = { workspace = true }
 serde = { workspace = true }
-spin-expressions = { path = "../expressions" }
 spin-factor-variables = { path = "../factor-variables" }
 spin-factor-wasi = { path = "../factor-wasi" }
 spin-factors = { path = "../factors" }
 spin-locked-app = { path = "../locked-app" }
 spin-manifest = { path = "../manifest" }
+spin-outbound-networking-config = { path = "../outbound-networking-config" }
 spin-serde = { path = "../serde" }
 tracing = { workspace = true }
 url = { workspace = true }
-urlencoding = "2"
 webpki-roots = "0.26"
 
 [dev-dependencies]
