+++ crates/loader/src/local.rs
@@ -4,7 +4,6 @@ use anyhow::{anyhow, bail, ensure, Context, Result};
 use futures::{future::try_join_all, StreamExt};
 use reqwest::Url;
 use spin_common::{paths::parent_dir, sloth, ui::quoted_path};
-use spin_factor_outbound_networking::SERVICE_CHAINING_DOMAIN_SUFFIX;
 use spin_locked_app::{
     locked::{
         self, ContentPath, ContentRef, LockedApp, LockedComponent, LockedComponentDependency,
@@ -13,6 +12,9 @@ use spin_locked_app::{
     values::{ValuesMap, ValuesMapBuilder},
 };
 use spin_manifest::schema::v2::{self, AppManifest, KebabId, WasiFilesMount};
+use spin_outbound_networking_config::allowed_hosts::{
+    AllowedHostsConfig, SERVICE_CHAINING_DOMAIN_SUFFIX,
+};
 use spin_serde::DependencyName;
 use std::collections::BTreeMap;
 use tokio::{io::AsyncWriteExt, sync::Semaphore};
@@ -142,7 +144,7 @@ impl LocalLoader {
         let allowed_outbound_hosts = component
             .normalized_allowed_outbound_hosts()
             .context("`allowed_http_hosts` is malformed")?;
-        spin_factor_outbound_networking::AllowedHostsConfig::validate(&allowed_outbound_hosts)
+        AllowedHostsConfig::validate(&allowed_outbound_hosts)
             .context("`allowed_outbound_hosts` is malformed")?;
 
         let component_requires_service_chaining = requires_service_chaining(&component);
@@ -838,7 +840,7 @@ fn requires_service_chaining(component: &spin_manifest::schema::v2::Component) -
 }
 
 fn is_chaining_host(pattern: &str) -> bool {
-    use spin_factor_outbound_networking::{AllowedHostConfig, HostConfig};
+    use spin_outbound_networking_config::allowed_hosts::{AllowedHostConfig, HostConfig};
 
     let Ok(allowed) = AllowedHostConfig::parse(pattern) else {
         return false;
