+++ crates/outbound-networking-config/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod allowed_hosts;
+pub mod blocked_networks;
